refactor(solicitacoes): refatorando a logica para pegar as informacoes da tabela por meio de serivce, mudando tambem para exclusao de fato
